Updated es_ES translation, based on version 4.1.4
ElPolloOscuro, part of my team (LokuraPlay) has completed the Saber-Factions Spanish translation. We have tested this on our server and it is only pending a general grammar review.

We'll update this in the future as we use the plugin, for now, this translation could help others and save a lot of time.

Feel free to update us as needed as well.